# Add an ability to query files in the current directory as displayed in lf