--- a/app.go
+++ b/app.go
@@ -552,6 +552,7 @@ func (app *app) runShell(s string, args []string, prefix string) {
 	gState.data["cmds"] = listCmds()
 	gState.data["jumps"] = listJumps(app.nav.jumpList, app.nav.jumpListInd)
 	gState.data["history"] = listHistory(app.cmdHistory)
+	gState.data["files"] = listFilesInCurrDir(app.nav)
 	gState.mutex.Unlock()
 
 	cmd := shellCommand(s, args)