### What it does

Implements https://github.com/gokcehan/lf/issues/1084 as `lf -remote "query $id files"`.

It outputs `\n` separated list of full file (dir/link/etc) paths of all files in lf's current directory visible at the moment with the current settings. (But it doesn't care about your scroll position)

The benefit compared to using external sorting tools is that `files` maintains the same sorting order and filter as used in lf at the moment.

This means that you can continue to view your files in some external program, and the "next" and "previous" will mean the same thing as in lf, opening the same files as "j" and "k" in lf.

### Example

Example usage with [nsxiv](https://codeberg.org/nsxiv/nsxiv) (image viewer):

<details><summary>lfrc open nsxiv</summary>
<p>

```bash
set shell bash
set shellopts '-eu'
set ifs "\n"

cmd on-init &{{
    # show id at bottom right corner
    rfmt="  %a|  %p|  \033[7;31m %m \033[0m|  \033[7;33m %c \033[0m|  \033[7;35m %s \033[0m|  \033[7;34m %f \033[0m|  %i/%t [id = $id]"
    lf -remote "send $id set rulerfmt \"$rfmt\""
}}

cmd open &{{
    case $(file -b -L --mime-type -- $f) in
        text/* | */xml | application/json)
            lf -remote "send $id editor $f"
            ;;
        image/*)
            images="$(lf -remote "query $id files" |
                grep -Ei "\.(png|jpg|jpeg|gif|webp|tif|tiff|bmp|heic|heif|avif|jxl|svg)(_large)*$")"
            num_current="$(echo "$images" | nl | grep -F -- "$(basename $f)" | awk '{print $1}')"
            echo "$images" | nsxiv -io -n $num_current 2>/dev/null | while read -r file; do
                [ -z "$file" ] && continue
                # select "marked" [press m in nsxiv] files
                lf -remote "send $id toggle \"$file\""
            done
            ;;
        *)
            setsid -f $OPENER $f >/dev/null
            ;;
    esac
}}

map l open
```

</p>
</details>

### The current state

So far, it works as expected, that is:

1. exports (read: updates `gState.data["files"]`) the correct file paths
2. maintains the correct order with the correct content (e.g. dironly, filter).
3. does it before you run shell commands, assuming the directory is done loading:
    * so it won't work is some cases in custom `on-...` commands, typically just after launch

Basically, it should look like what you see in lf, if you scrolled all the way and copied everything, and resolved full path for everything. If it doesn't, it's a bug.

#### TODO - this needs fixing or at least being addressed

- [ ] when `set dirpreviews true` the current directory isn't done `loading` until the previewer script exits, even though it runs on directory's **child**, thus shouldn't really block it. And it walks all dirs from / to the viewed, increasing the delay. The patch from https://github.com/gokcehan/lf/pull/1949#issuecomment-2785207399 seems to fix it, switching to the existing regular file previewer function.

### Implementation

<details><summary>Old</summary>
<p>

It works by calling `app.nav.exportFilesInCurrDir()` in a few key places, which updates `gState.data["files"]` if `nav` initialized and the current directory is done loading.

For convenience, `exportFilesInCurrDir` takes `origin` of call as the argument that you can pass to logs, but I don't expect we keep it (too noisy).

To choose the places of calls, I've gone over references of [`dir.sort()`](https://github.com/gokcehan/lf/blob/6718a363349c5958acc743c261a6f241c33c3401/nav.go#L211), avoiding exporting right after, and instead trying to find some higher level function where possible to avoid needless exports.

It seems like I could _almost_ solely rely on export where `dirChan` is consumed. I explicitly added some exports just in case there are some corner cases I didn't account for, but it seems like `rename`, `onChdir`, `nav.sort` are covered by `loop.case app.nav.dirChan` export.

The only thing it doesn't cover is `setFilter`, which I added too.

_Original implementation mentioned the above that always maintains `files` in sync with what is displayed in lf can be found here:_

https://github.com/MahouShoujoMivutilde/lf/tree/query-files-always-in-sync

(force-pushed out of this branch to not clutter history)

</p>
</details> 

Replaced with the proposed export in `runShell()`, as it works fine, and when it has problems (`set dirpreviews true`) - mine does too.

---

So, this is where we're at. It works fine for me. At this point it will be more productive to work on it in pr, should more changes be needed.