@MahouShoujoMivutilde 2025-04-06
Thanks for the feedback!

Yes, I understand. Your implementation is the first thing I tried after looking how the other `query` variables are saved, but there were some issues that forced me to abandon it.

After some testing, I see that it panics in `on-cd` on start:

```bash
cmd on-cd &{{
    echo noop
}}

# this causes the panic
on-cd
```


<details><summary>Details</summary>
<p>

```log
panic: runtime error: index out of range [-1]

goroutine 1 [running]:
main.(*nav).currDir(...)
/home/witch/projects/forked/elefu-his/nav.go:1982
main.(*app).runShell(0xc0000c8dd0, {0xc00067c000, 0x2b8}, {0x0, 0x0, 0x0}, {0xc0006766c0, 0x1})
/home/witch/projects/forked/elefu-his/app.go:555 +0xca8
main.(*execExpr).eval(0xc000270740, 0xc0000c8dd0, {0x0, 0x0, 0x0})
/home/witch/projects/forked/elefu-his/eval.go:2250 +0x3b9
main.(*callExpr).eval(0xc000678180, 0xc0000c8dd0, {0x0?, 0xc000000000?, 0x4cdf27?})
/home/witch/projects/forked/elefu-his/eval.go:2233 +0x7e5f
main.(*app).readFile(0xc0000c8dd0, {0xc0000183a0, 0x1b})
/home/witch/projects/forked/elefu-his/app.go:110 +0x194
main.(*app).loop(0xc0000c8dd0)
/home/witch/projects/forked/elefu-his/app.go:279 +0x1746
main.run()
/home/witch/projects/forked/elefu-his/client.go:65 +0x357
main.main()
/home/witch/projects/forked/elefu-his/main.go:357 +0x819
```


</p>
</details>

So at it should include the same guards my export has.


<details><summary>E.g. like this</summary>
<p>

```patch
diff --git a/app.go b/app.go
index eaa913c..a383bbb 100644
--- a/app.go
+++ b/app.go
@@ -552,6 +552,8 @@ func (app *app) runShell(s string, args []string, prefix string) {
 	gState.data["cmds"] = listCmds().String()
 	gState.data["jumps"] = listJumps(app.nav.jumpList, app.nav.jumpListInd).String()
 	gState.data["history"] = listHistory(app.cmdHistory).String()
+	// gState.data["files"] = listFiles(app.nav.currDir()).String()
+	app.nav.exportFilesInCurrDir()
 	gState.mutex.Unlock()

 	cmd := shellCommand(s, args)
diff --git a/nav.go b/nav.go
index 0935ff0..5baffac 100644
--- a/nav.go
+++ b/nav.go
@@ -718,6 +718,18 @@ func (nav *nav) exportFiles() {
 	}
 }

+// Updates gState.data["files"], doesn't lock
+func (nav *nav) exportFilesInCurrDir() {
+	if !nav.init {
+		return
+	}
+	dir := nav.currDir()
+	if dir.loading {
+		return
+	}
+	gState.data["files"] = listFiles(dir).String()
+}
+
 func (nav *nav) dirPreviewLoop(ui *ui) {
 	var prevPath string
 	for dir := range nav.dirPreviewChan {
diff --git a/ui.go b/ui.go
index 54dd103..67c48a9 100644
--- a/ui.go
+++ b/ui.go
@@ -1204,6 +1204,16 @@ func listHistory(history []cmdItem) *bytes.Buffer {
 	return b
 }

+func listFiles(dir *dir) *strings.Builder {
+	b := new(strings.Builder)
+
+	for _, file := range dir.files {
+		fmt.Fprintln(b, file.path)
+	}
+
+	return b
+}
+
 func listMarks(marks map[string]string) *bytes.Buffer {
 	t := new(tabwriter.Writer)
 	b := new(bytes.Buffer)

```

</p>
</details>

I don't remember precisely where else it fails, but I remember there was _something_, which is why I went with backtracking from `dir.sort()` calls method. (I was struggling to get it to work consistently, didn't exactly write everything down)

The basic testing shows that it works, including the filter which I initially suspected would fail.

I'll see if I can spot anything else.

---

In regards to the added latency, it seems fine.

* for the directory with 6500 photos - `exportFilesInCurrDir()` took 4.632001ms; all `runShell()` exports took 5.755784ms (includes files). 
* For 50 files it's `exportFilesInCurrDir()` took 14.76µs; all `runShell()` exports took 575.982µs.


<details><summary>Details</summary>
<p>

```patch

diff --git a/app.go b/app.go
index eaa913c..9657948 100644
--- a/app.go
+++ b/app.go
@@ -541,6 +541,7 @@ func (app *app) runCmdSync(cmd *exec.Cmd, pause_after bool) {
 //	!       Yes   No     Yes    Yes     Yes     Pause and then resume
 //	&       No    Yes    No     No      No      Do nothing
 func (app *app) runShell(s string, args []string, prefix string) {
+	startRS := time.Now()
 	app.nav.exportFiles()
 	app.ui.exportSizes()
 	app.ui.exportMode()
@@ -552,6 +553,10 @@ func (app *app) runShell(s string, args []string, prefix string) {
 	gState.data["cmds"] = listCmds().String()
 	gState.data["jumps"] = listJumps(app.nav.jumpList, app.nav.jumpListInd).String()
 	gState.data["history"] = listHistory(app.cmdHistory).String()
+	startFiles := time.Now()
+	app.nav.exportFilesInCurrDir()
+	log.Printf("exportFilesInCurrDir() took %s", time.Since(startFiles))
+	log.Printf("all runShell() exports took %s", time.Since(startRS))
 	gState.mutex.Unlock()

 	cmd := shellCommand(s, args)
diff --git a/nav.go b/nav.go
index 0935ff0..5baffac 100644
--- a/nav.go
+++ b/nav.go
@@ -718,6 +718,18 @@ func (nav *nav) exportFiles() {
 	}
 }

+// Updates gState.data["files"], doesn't lock
+func (nav *nav) exportFilesInCurrDir() {
+	if !nav.init {
+		return
+	}
+	dir := nav.currDir()
+	if dir.loading {
+		return
+	}
+	gState.data["files"] = listFiles(dir).String()
+}
+
 func (nav *nav) dirPreviewLoop(ui *ui) {
 	var prevPath string
 	for dir := range nav.dirPreviewChan {
diff --git a/ui.go b/ui.go
index 54dd103..67c48a9 100644
--- a/ui.go
+++ b/ui.go
@@ -1204,6 +1204,16 @@ func listHistory(history []cmdItem) *bytes.Buffer {
 	return b
 }

+func listFiles(dir *dir) *strings.Builder {
+	b := new(strings.Builder)
+
+	for _, file := range dir.files {
+		fmt.Fprintln(b, file.path)
+	}
+
+	return b
+}
+
 func listMarks(marks map[string]string) *bytes.Buffer {
 	t := new(tabwriter.Writer)
 	b := new(bytes.Buffer)

```

</p>
</details>