@joelim-work 2025-04-07
@MahouShoujoMivutilde Good catch with finding the panic, I forgot that functions like `app.nav.currDir()` assume that `lf` has already been initialized. I think it's fine to keep all the code in a single function, just change the structure so it's like this:

```go
func listFiles(nav *nav) string {
	// return "" if there's an error
}
```

Also just out of interest, what are the implications if the directory is `loading`? Would it be better to return nothing at all, or just proceed anyway and return potentially outdated information?